Meteor-Hello-World
Ran into a series of problems with meteor create. Found someone's solution here: https://github.com/meteor/meteor/issues/8521 but it did not work and instead prevented me from applying Philip's fix posted on slack. I ended up deleting the contents of repo, to apply the fix and 'meteor npm install' worked, but not meteor run. This is had something to do with my repo not connecting with git. My repo wouldn't connect with Github and I was unable to commit, and unable to reclone so I deleted everything and started again. Worked smoothly this time and everything works